Code refactor : Removed config xml, make use of sequenceDao for geenrating sequence

+++ bundles/cwf-dbhelper/src/main/java/cwf/dbhelper/sequencegenerator/SequenceDao.java
@@ -3,5 +3,5 @@ package cwf.dbhelper.sequencegenerator;
 import cwf.dbhelper.sequencegenerator.SequenceException;
 
 public interface SequenceDao {
-	long getNextSequenceId(String key) throws SequenceException;
+	int getNextSequenceId(String key) throws SequenceException;
 }
